Restore catch as a real middleware
Get rid of the separate built_in sublibrary along the way, as it was
becoming an obstacle in the dependency graph.

The catch middleware was originally deleted as a separate middleware,
and eventually replaced by an internal middleware during error handler
unification. However, the plan is now to make that internal middleware
customizable, so restore it again as a middleware. It's still built-in,
but only as a default that can be changed.