CI: test only on OCaml 4.13 and higher
Dream no longer installs on lower OCaml versions, because mirage-crypto
requires OCaml 4.13. See

  https://github.com/mirage/mirage-crypto/pull/233

The reason for mirage-crypto requiring 4.13 seems to be minor
convenience, and isn't justified in my opinion. However, Dream itself
will likely require OCaml 5 in the medium term due to upgrading to
multicore and effects, so it seems unnecessary to make an effort to
restore 4.08 support at this point.